+++ src/main/java/com/project/config/NotificationsConfig.java
@@ -1,0 +1,13 @@
+package com.project.config;
+
+public class NotificationsConfig {
+    private TelegramConfig telegram = new TelegramConfig();
+
+    public TelegramConfig getTelegram() {
+        return telegram;
+    }
+
+    public void setTelegram(TelegramConfig telegram) {
+        this.telegram = telegram;
+    }
+}